+++ src/layers/DialogData.hpp
@@ -1,0 +1,19 @@
+const std::vector<std::pair<std::string, int>> ScratchDialog = {
+    //  REFERENCE
+    //  { Dialog, Character Pic }
+
+    {"Oh hey...<d020> It's <cg>you</c> again", 12},
+    {"Back at this ol' spot?<d020> Remember you emptied it.", 13},
+    {"It's surprising that you had enough currency to buy everything from <cy>all the shops</c> in the game.", 26},
+    {"I know that <cp>Purple Bot</c> doesn't sell stuff for cheap, even after a nerf.", 13},
+    {"<cb>100,000 orbs</c> for a music customizer?<d020> At that point it's better spending it on the other shops.", 12},
+    {"Well, except the community one since that one didn't get a restock for years.", 8},
+    {"Maybe in a <cy>future update</c>, perhaps?", 9},
+    {"There's still around <cr>200 Icons</c> locked until 2.21...", 11},
+    {"You better start gathering more orbs,<d010> diamonds<d010> and moons...", 13},
+    {"Maybe it would be better for you if Rub added more achievements based on level difficulties.", 8},
+    {"Well...<d020> Except demon-based ones.<d020> Who would be insane enough to grind 500 Easy demons?", 26},
+    {"Oh yeah...<d020> The <cc>guy</c> who made this <cg>mod</c>.", 12},
+    {"Anyway, it's nice for you to stop by, but I have nothing to offer...", 13},
+    {"Gonna wait for another restock, hopefully soon...", 8}
+};